docs: comment-policy sweep across feature/sync-server + extract me-fetcher
Drift bug fixes:
- migrations/0001 referenced 0002_create_assets.sql (renamed to
  0002_create_tldraw_assets.sql).
- migrations/0001 pinned the convertLocalDocToSynced function name
  in two created_at-related comments.
- app/migration.ts pinned packages/worker/src/assets.ts (renamed
  to tldraw-assets.ts).

Worker comment trims: the SOT-mirror "Ensure the SQLite snapshot
table exists" preamble in DocumentSyncRoom; the flushSnapshot JSDoc
that just restated the method; the peekSnapshotVersion JSDoc usage
description ("used by the initializing-doc cleanup sweep"); the
duplicating second JSDoc in tldraw-asset-policy plus its
exports["./asset-policy"] filename pin; the wrangler.toml "10 min,
see cleanup.ts" specific pin and the derived "~15 minutes"
arithmetic.

App comment trims: the migration filename pin in api-repository.ts
plus the save-method JSDoc that duplicated the DocumentRepository
interface; the file-top preamble in useSyncedRepository.ts that
duplicated the hook JSDoc; three section-divider preambles, one
restate-merge-order comment, and one misplaced JSDoc in
useDocumentManager.ts; the uncited "6 per host" HTTP/1.1 cap claim
and the "10 MB payload" SOT mirror in migration.ts; two
restate-next-call comments in reconnect.ts; the long error-string
enumeration JSDoc in conversion-error-message.ts.

me-fetcher extraction: move fetchMe out of AuthContext.tsx into a
sibling me-fetcher.ts so AuthContext.tsx is "components only" and
HMR partial-reload works; resolves the
react-refresh/only-export-components lint warning.
