+++ packages/worker/wrangler.toml
@@ -3,14 +3,12 @@ main = "src/worker.ts"
 compatibility_date = "2025-05-08"
 compatibility_flags = ["nodejs_compat"]
 
-# Scheduled trigger for the initializing-doc sweep. A doc creation is
-# a multi-step client flow (POST → asset upload → snapshot push); the
-# server marks rows `initializing_at = now()` on POST and clears it
-# on the finalizing snapshot push. Rows that linger past the grace
-# window (10 min, see cleanup.ts) are abandoned creates from a tab
-# close, crash, or delete-mid-migration; this cron sweeps them out.
-# Five-minute cadence so the worst-case visibility for an abandoned
-# row is ~15 minutes.
+# Scheduled trigger for the initializing-doc sweep. A doc creation
+# is a multi-step client flow (POST → asset upload → snapshot push);
+# the server marks rows `initializing_at = now()` on POST and clears
+# it on the finalizing snapshot push. Rows that linger past the
+# grace window are abandoned creates from a tab close, crash, or
+# delete-mid-migration; this cron sweeps them out.
 [triggers]
 crons = ["*/5 * * * *"]
